pyproject.toml
[project]
name = "runpod-serverless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "datasets>=4.4.1",
    "fastapi>=0.129.0",
    "gradio>=6.5.1",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "kokoro>=0.9.4",
    "librosa>=0.11.0",
    "locust>=2.43.3",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "piper-tts>=1.3.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "soundfile>=0.13.1",
    "torchcodec>=0.9.0",
    "uvicorn>=0.40.0",
]
